backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
sqlite-utils
pandas
python-dateutil
cors
pymysql
python-dotenv
pdfplumber
psycopg2-binary
gunicorn
